expose single-query batching executor
`@graphql-tools/batch-execute` exposes most of what we need for this - an
executor that calls through a DataLoader with a load function that
handles the many-to-one merging of outbound requests it batches
together. What it _doesn't_ do - that we _need_ - is expose the ability
to further divide the batches produced by the DataLoader - in our case,
to divide them into groups of queries that are safe to execute together.

Because this is a hard requirement and we don't want to copy & rewrite
the entire package in order to plumb our extra batching logic to the
dataloader, we instead use yarn to patch the package so it _does_ expose
the functionality, more or less. Most of the logic is in the
[loader](https://github.com/ardatan/graphql-tools/blob/a648a9371c0dfe77df0e8afce3eb569dafa947d1/packages/batch-execute/src/createBatchingExecutor.ts),
so we export the loader, allowing us to wrap it however we please, in
this case with our pre-existing `subBatch` method.