Initial configuration managed by @taql/config
Anything that wants to access process.env must go through there. This is
mostly to ensure that it's easy to discover all of the configuration
available and onerous to require too much configuration - that is, it
encourages our configuration to be simple and easy to understand. Later,
it can also provide a central location to hook in something better than
the `dotenv` package to manage overrides for dev environments.