Add koa - we are gonna need middleware
Instead of using graphql-yoga directly as a server, we will use koa to
handle the web stack & lifecycle. It's small, simple, well-documented
and well-supported. In the future, we'll be dealing with input
transformations to support endpoints exposing slightly different APIs
(queries in an array on an object in the request, instead of the array
being the request itself, that sort of thing), and the easiest way to
manage that won't be with complicated envelop plugins to reshape the
request once yoga has its hands on it: instead we can write simple koa
middleware to reshape the request before forwarding it along.

This enables us to move header management outside of envelop and into
koa's middleware, which makes more sense given headers fit better on
a per-request lifecycle (middleware) than on a potentially per-query
lifecycle (envelop).

We also sanitize state management a bit, introducing the idea of a
generic taql plugin that can offer a mix of envelop plugins and koa
middlewares to the server.