Don't forward cookie or user-agent header(s)
I was afraid legacy graphql or some stitched service upstream of it
consumed cookies, but based on a close read of graphql and our service
call internals, I am inclined to think I was wrong. Given that letting
services consume cookies is obviously a bad idea (if we want to pass
user state, let it be in well-standardized headers app-proxy extracts
for us), I'm pulling it for now. With any luck, actual testing later
won't show cookies to be important.

As far as user agent, 'forwarding' that just overwrites our outbound UA.
Obviously not the best idea. The client's UA is already passed on
`x-ua`, and I _hope_ that's only used for reporting purposes by
upstreams.