Config type changes
Config resolution enforces that when a default type is provided without
a resolver and the type of the default is _not_ string, that the result
type of resolving the configuration is string & the default type, not
the default type by itself. That is what the code _does_, so it is worth
making sure the type system knows.

Also overhauled resolver error handling. Why reinvent the exception when
you can throw an exception?