Overhaul shared context, get rid of plugin model
The existing plugin model was making it hard to mix our plugins (typed
against the context we know we build) with arbitrary plugins, and there
are easier ways to deal with it. Instead of using generics to drive
distribution of our context through yoga, koa, and envelop plugins, we
use a constant context initialized early in the request flow. Plugins
that used generics to extend the context to add private state (e.g., the
batching system) now use WeakMaps keyed on the context to hold their
state and keep it private. This has the added benefit of completely
preventing key collisions in the state object.

All things that were envelop plugins are now yoga plugins, because yoga
plugins are just envelop plugins with a few extra lifecycle hooks.

I fiddled with plugin order a bit - The concurrency plugin in particular
was added too late (last) and as such nothing was happening between
incrementing and decrementing the concurrency counter.

This change has a surreptitious extra goal - by simplifying the context,
it becomes serializable and we have a shot at passing it to worker
threads to respond to queries.