remove erroneous ttls from lru-caches
Caches provided by `lru-cache` do not automatically purge stale entries
by default. The `ttl` instead is used to detect stale entries on
retrieval. In line with this, for caches with values that cannot
meaningfully become stale (i.e., the validity of a document: the store
is associated with the schema so if the document doesn't change - it
can't, it's the key - the validity doesn't change), we remove the ttls.
There's no point attempting to use ttls to free memory (by configuring
autopurging, or periodically purging manually using `setInterval`, because we must
have enough memory available to allow the cache to be full, and if we
don't expect the cache to be full at runtime, we should configure it to
be smaller. (Presumably this reasoning is why lru-cache doesn't enable
autopurging by default)