Replace generic caching plugins with unified caching plugin
The document cache sees the most benefits: We preferentially use
preregistered query ids and persisted query ids as cache keys, so the
raw query string need not be hashed - a small savings on small queries
and a surprisingly large savings on large queries.

The error cache is combined with the document cache, and is no longer
needed. Other caches (jit, validation, etc) used to layer a WeakMap on
the parsed DocumentNodes with a (provided) cache on the query string -
we always provided instrumented LRU caches. This makes sense for generic
plugins, but we know that the DocumentNode will almost always have come
from a hit on our parser caches, so we can rely entirely on the WeakMap
and do away with the secondary cache. Additionally, our jit cache
compiles documents lazily instead of blocking the active request to
complete compilation.

Finally, we prewarm all caches (document, validation, jit) based on the
current contents of the preregistered query database and the persisted
query redis instance.