Security plugin may prevent arbitrary operation execution
When taql is configured with RUN_UNTRUSTED_OPERATIONS=false and an
inbound request has the 'x-taql-trust-client: false' header set, taql
will refuse to run arbitrary queries, instead requiring preregistered or
persisted queries to be used. As long as taql remains behind a proxy
that doesn't forward arbitrary headers (in this case, TRTOP), this is
good enough. We can move to more robust workload authentication (oidc,
etc) later.

This patch presents an unfortunate choice:

To minimize processing done on untrusted queries, we must check as soon
as possible whether they are trustworthy. An untrusted client shouldn't
so much as trigger a redis or database lookup! Because the preregistered
query plugin works within the `onParams` phase, we must execute in that
phase before the preregistered query plugin runs, or as a koa middleware
before the request is fed to yoga. The `onParams` phase is not given
access to the koa context (though it must be trivially available),
probably to preserve the function of envelop plugins  using the
`onEnveloped` callbacks (which assume they have complete control over
the context). This forces an unfortunate choice: we must either leak
header management out of the @taql/context package to expose the header
to the security plugin, or we must leak graphql implementation details
out of yoga entirely and inspect the request (for preregistered query
ids, etc) in middleware. Neither is particularly palatable, but the
latter felt worse, so for the time being I'm letting header management
leak.