Prereg plugin updates stores
The prereg plugin tries to be nice to the database when it refreshes the
allowlist of preregistered queries by requesting only the keys added
since the plugin last checked. The database assumes our threshold is
_seconds_ since the Unix epoch. The plugin has been providing a
threshold in _milliseconds_ since the Unix epoch. Some hasty math
suggests that as configured, we would first update the preregistered
queries store in ~53,000 years, and a second time in
~53,000,000 years (should a taql instance be left running long enough).
This is unsatisfactory, so instead we make sure we send a measure in
_seconds_ to the database, and ask for reasonable things like
queries added since 1687294095 seconds post epoch instead of
1687294095000 seconds post epoch.

The plugin also refreshes less aggressively - it used to try (and fail)
every 10 seconds. Now we will try every 30 seconds. If not updating at
all was good enough for us to get this far without noticing, I'm sure
waiting an additional 30 seconds will do us just fine.

Also, I extended the prereg plugin labels to include the worker in
question.