taql now logs errors we return to clients in its own logs
For example, the 'view map' link isn't working on 'Things to Do Near'
pages. Before, the best way to get the relevant error message was to
inspect the results of the graphql call in the browser. Now, if the
result is an error message, we find this in our logs:

```
error: Variable "$locationIds" got invalid value "32655" at "locationIds[0]"; Int cannot represent non-integer value: "32655"

GraphQL request:1:19
1 | query FetchLatLng($locationIds: [Int!]!) {
  |                   ^
2 |   locations(locationIds: $locationIds) { {"worker":"1"}
```

Presumably we'll now see other related issues without waiting for them
to be reported and struggling to reproduce them.