SVCO handling controlled by schema
Why try to make the context guess whether or not to handle any given
svco? The schema has a better sense of what is relevant to it (e.g., the
host of the legacy subgraph), so we let it be in control.

I also noticed some unsoundness around the return type of makeSchema (it
could not tell a partial success from an error at all in the type
system), so I made the success/partial-success/error states explicit on
the return type.