useTaqlSecurity: prepare for /data/graphql DTM
As described in
https://gitlab.dev.tripadvisor.com/tripadvisor/web/-/merge_requests/15917,
we need to tolerate `preRegisteredQueryId` being redundantly present in
`query` for untrusted traffic.

It's actually already in both `query` and `extensions` today, which is
okay because the implementation of /data/graphql/ids in TRTOP marks its
requests as trusted. (And /data/graphql/batched is marked as untrusted.)

But that setup will no longer with DTM, as app-proxy does not parse
request bodies; it will instead blindly proxy /data/graphql/ids and
/data/graphql/batched onto taql's /graphql, streaming the request
bodies. So, app-proxy will not be able to mark traffic as trusted or not
trusted based on the request path, as the semantic distinctions between
those paths will disappear: clients will be able to send pre-registered
queries to /batched or normal queries to /ids. So, we have no choice but
to mark all prod traffic as untrusted. So, we need to make this change.