Bypass graphql-yoga's schema caching plugin
It's buggy! If we have to use the yoga server's `handleNodeRequest`
method and dynamic schemas, there's no way to win with the built-in
schema caching plugin. The schema plugin uses a weakmap on the request
object at the _parse stage_ to associate the schema with the request,
but it uses a weakmap on _a_ request object in the context when it's
time to recover the schema. Notably, there is _no requirement that these
are the same object_. In fact, when using `handleNodeRequest`, it is
all-but guaranteed they are not, because this method normalizes the
request object, replacing it with a _new object_ in the call stack (but
doesn't touch the context). There might be a way to totally rework our
context to accomplish this, but using our own schema plugin is good
enough for now. At the end of the day, the problem smells like a _bug_
so changing our context to work with it sounds wrong, compared to
bypassing buggy code.