Use generic header handling for logging plugin
The logging plugin (which logs errors) assumes the wrong type of headers
object and calls request.headers.get(...), then blows up.

One thing we don't do well in taql is manage the context. The interplay
between koa, envelop, and yoga is complicated to say the least, and we
know where to go to get the data taql adds (practially by trial and
error)... the intrinsic data, like the originating request, or whatever
transformation or normalization thereof any given layer of the
application is interested in, is a little harder to stay on top of. That
is to say, several types of object may be in `context.request` in any
given place in the code, and with them several different types of header
objects.

After the last round of updates to koa/envelop/yoga, this situation got
worse. While we haven't had time to dig in and get it all sorted out,
we've already come up with a way to handle the three distinct header
formats we've encountered in request objects: Our own headers package,
which discovers what type of headers object it has before trying to read
the headers themselves, instead of counting on the type system (which,
unfortunately, tends to be wrong about what type of request and headers
will be present in the runtime). This should fix the error reporting
bug. With any luck, the errors that will now be successfully reported
won't be errors I have to fix, too.