Work around envelop prometheus plugin
A developer reports:
> I'm currently seeing `[GraphQL] Added label "phase" is not included in
> initial labelset: [ 'operationName', 'operationType', 'errorPhase' ]`
> when trying to override WPS on webml

The envelop prometheus plugin is inconsistent with the labels it passes
to the error counter. It used to _think_ it always passed `phase` (by
calling the `fillLabelsFn`, which by default passed `phase`) but
occasionally failed to call the label generation function and passed
`errorPhase` instead. `prom-client` is opinionated and throws errors
when it finds labels it doesn't expect, so this turns mundane error
reporting (maybe a timeout, maybe a validation error) into an
exceptional runtime error. Not great! To get around this, last time, we
overrode the `fillLabelsFn` such that it always convered `phase` to
`errorPhase`, to match the _bad_ calls that bypassed `fillLabelsFn`.

With the latest `envelop` upgrade, they've _fixed_ these calls, not by
_not_ bypassing `fillLabelsFn`, mind you - that would be too hard and
we're not dealing with geniuses here - but by at least passing the
more-likely-to-be-consistent-with-`fillLabelsFn` `phase` label in the
cases that they bypass the `fillLabelsFn`.

Needless to say, that breaks our earlier workaround.

We do two things to resolve this:
1. We undo our earlier workaround: Now we pass the `phase` label. I've
   checked our graphs and alerts and none of them use the metric in
   question, so we're OK to change the label. Plus, `phase` is less
   redundant than `errorPhase` in a counter named for errors.
2. In case that wasn't good enough, we also implement a tricky method
   called `defang` to dynamically extend any class such that the
   extension won't throw errors from a selected subset of its void
   methods. We `defang` all the prom metrics we pass to envelop's
   `usePrometheus`, so if it gets the labels wrong, we'll see error logs
   in `taql` but not runtime errors failing requests completely.